Move type tests to __typetests__
