+++ src/__typetests__/__fixtures__/index.ts
@@ -1,0 +1,1 @@
+export * from './type';
