+++ src/__typetests__/types/__fixtures__/index.ts
@@ -1,4 +1,3 @@
 export * from './pathString';
 export * from './traversable';
 export * from './tuple';
-export * from './type';
