+++ src/__typetests__/types/__fixtures__/pathString.ts
