+++ src/__typetests__/types/__fixtures__/traversable.ts
