+++ src/__typetests__/types/__fixtures__/tuple.ts
